Refactored file names to all use underscore_case
